{Content->Download}Manager::getDownloadInfo()
